gh-154 Standartize Button Types - add const ButtonTypes - Move const ImageShapes into Image
