--- apps/web-portal-client/lib/components/ImageShapes.jsx
@@ -1,4 +1,0 @@
-export const ImageShapes = {
-  cross: 'M6 18L18 6M6 6l12 12',
-  sandwich: 'M4 6h16M4 12h16M4 18h16',
-}
