+++ apps/web-portal-client/lib/components/index.jsx
@@ -1,4 +1,3 @@
 export * from './HomePage'
 export * from './Button'
 export * from './Image'
-export * from './ImageShapes'
